requirements.txt
django~=5.1.3
pillow~=11.0.0
djangorestframework~=3.15.2
psycopg2~=2.9.10
django-ckeditor~=6.7.1
django-allauth~=65.2.0
requests~=2.32.3
python-dotenv~=1.0.1